Split MultistreamProtocol interfaces and implementations
